MdeModulePkg/PciSioSerialDxe:bitwise operation have same width operands
Operands in a bitwise operation have different size. Update code to fix it.

(Sync patch r19309 from main trunk.)

Contributed-under: TianoCore Contribution Agreement 1.0

git-svn-id: https://svn.code.sf.net/p/edk2/code/branches/UDK2015@19323 6f19259b-4bc3-4df7-8a09-765794883524

+++ MdeModulePkg/Bus/Pci/PciSioSerialDxe/Serial.c
@@ -982,7 +982,7 @@ SerialControllerDriverStart (
             &Supports
             );
           if (!EFI_ERROR (Status)) {
-            Supports &= EFI_PCI_IO_ATTRIBUTE_IO | EFI_PCI_IO_ATTRIBUTE_MEMORY;
+            Supports &= (UINT64)(EFI_PCI_IO_ATTRIBUTE_IO | EFI_PCI_IO_ATTRIBUTE_MEMORY);
             Status = ParentIo.PciIo->Attributes (
               ParentIo.PciIo,
               EfiPciIoAttributeOperationEnable,
